+++ plugins/nv-filters/data/locale/en-US.ini
@@ -14,3 +14,6 @@ Nvvfx.Method.Greenscreen.Threshold="Threshold"
 Nvvfx.OutdatedSDK="WARNING: Please upgrade both NVIDIA Video & Audio SDK. Your current version of Video SDK is outdated."
 Nvvfx.Method.Greenscreen.Processing="Mask refresh frequency in frames"
 Nvvfx.Method.Greenscreen.Processing.Hint="This alleviates GPU load by generating a mask every N frames only (2 on default)."
+Nvvfx.Method.BlurFilter="NVIDIA Blur Filter"
+Nvvfx.Method.BackgroundBlurFilter="NVIDIA Background Blur Filter"
+Nvvfx.Method.Blur.Strength="Blur Intensity"
