+++ plugins/nv-filters/nv-filters.c
@@ -14,6 +14,8 @@ extern void unload_nvidia_afx(void);
 #endif
 #ifdef LIBNVVFX_ENABLED
 extern struct obs_source_info nvidia_greenscreen_filter_info;
+extern struct obs_source_info nvidia_blur_filter_info;
+extern struct obs_source_info nvidia_background_blur_filter_info;
 extern bool load_nvidia_vfx(void);
 extern void unload_nvidia_vfx(void);
 #endif
@@ -29,8 +31,11 @@ bool obs_module_load(void)
 	obs_enter_graphics();
 	const bool direct3d = gs_get_device_type() == GS_DEVICE_DIRECT3D_11;
 	obs_leave_graphics();
-	if (direct3d && load_nvidia_vfx())
+	if (direct3d && load_nvidia_vfx()) {
 		obs_register_source(&nvidia_greenscreen_filter_info);
+		obs_register_source(&nvidia_blur_filter_info);
+		obs_register_source(&nvidia_background_blur_filter_info);
+	}
 #endif
 	return true;
 }
